Refactoring out
Pequeño refactoring aout
