+++ service/cmd/service/main.go
@@ -6,8 +6,8 @@ import (
 	"runtime"
 	"time"
 
-	"service/internal/broker"
 	"service/internal/conf/v1"
+	"service/internal/out/broker"
 	mylog "service/pkg/logger"
 
 	klog "github.com/go-kratos/kratos/v2/log"
